Moved API requests to own package
