+++ app/com/gu/identity/frontend/services/IdentityServiceRequestHandler.scala
@@ -2,6 +2,7 @@ package com.gu.identity.frontend.services
 
 import com.gu.identity.frontend.logging.{Logging => ApplicationLogging}
 import com.gu.identity.service.client._
+import com.gu.identity.service.client.request._
 import play.api.libs.json.Json
 import play.api.libs.json.Reads.jodaDateReads
 import play.api.libs.ws.{WSResponse, WSClient}
@@ -51,7 +52,7 @@ class IdentityServiceRequestHandler (ws: WSClient) extends IdentityClientRequest
         }
 
   def handleRequestBody(body: ApiRequestBody): String = body match {
-    case b:RegisterRequestBody => Json.stringify(Json.toJson(b))
+    case b: RegisterRequestBody => Json.stringify(Json.toJson(b))
     case AuthenticateCookiesRequestBody(email, password) => encodeBody("email" -> email, "password" -> password)
   }
